Excercise 4.21* Finished

+++ part4/Blog list/controllers/blog.js
@@ -80,8 +80,25 @@ blogsRouter.put('/:id', async (request, response, next) => {
 //  DELETE
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if ( blog.user.toString() === decodedToken.id.toString() ){
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    }else{
+      return response
+      .status(400)
+      .send({
+        error: 'You can only delete your own blogs'
+      })
+    }
+    
   } catch (exception) {
     next(exception)
   }
